Actualización

+++ ejercicios/3. Interracción con el usuario/interaccion.cpp
@@ -5,10 +5,9 @@ int main()
 {
     string nombre = "Juan";
     int edad = 20;
-    int salario;
+    float salario;
     cout << "Introduzca el salario del empleado: " << endl;
     cin >> salario;
-
     cout << "El trabajador " << nombre << ", con la edad de " << edad << " anos, tiene un salario de " << salario << " Euros mensuales.";
     
     return 0;
